requirements.txt
python-dotenv==1.2.1
ollama==0.6.1
langchain==1.2.7
langchain-ollama==1.0.1
langchain-openai==1.1.7
